changed model
