+++ TableSettingsMVC/Scenes/App/AppDelegate.swift
