--- TableSettingsMVC/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  TableSettingsMVC
-//
-//  Created by Artyom Guzenko on 26.04.2022.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
